refactored compact filters
